+++ components/backup-plugins/before-commit.sh
